requirements.txt
tqdm
streamlit
langchain
langchain-community
langchain-groq
chromadb
python-dotenv
guardrails-ai
docx2txt
huggingface_hub
langchain-huggingface
faiss-cpu

